did a TON of work to battle system
i like small numbers in rpgs, so i'm basing the calculations around them (e.g. paper mario).
everything's basically be reworked to now be based around your base damage, which can be upgraded as the game goes along, and by your weapon damage. these two will be added together, with the enemy's defense subtracted from this, to form the basis of damage calculations.
also, i'm going to be porting everything over from the old project, to this one, so this can be the main hub from here on out.