new maps, battle entry animation
starting to build out the actual game a bit, now that i have basic systems in place.
still need to finish the menu (including tile swapping for the money count in the UI), and also figure out how to do an actual usable inventory, so the stick in the beginning actually does something.