started a crafting menu!
not much to it now besides some lil floating grape, wheat, and barley sprites. yay.
the idea is that you can craft "healing potions" in the form of "shedeh", a kind of proto-wine they used to make in ancient Egypt, and possibly beer, which is also a thing they had back then, which will act as an AP-recovery potion.